refactor(session): move cookie-specific files under cookie/ submodule
Groups config, extractor, and middleware into src/auth/session/cookie/
with a new mod.rs. Updates all super:: imports to use crate-absolute
paths. Back-compat re-exports in session/mod.rs keep external callers
unchanged. store.rs updated to import SessionConfig via cookie module.
