+++ src/auth/session/cookie/middleware.rs
@@ -14,9 +14,9 @@ use crate::cookie::{CookieConfig, Key};
 use crate::ip::ClientIp;
 
 use super::extractor::{SessionAction, SessionState};
-use super::meta::{SessionMeta, header_str};
-use super::store::SessionStore;
-use super::token::SessionToken;
+use crate::auth::session::meta::{SessionMeta, header_str};
+use crate::auth::session::store::SessionStore;
+use crate::auth::session::token::SessionToken;
 
 // --- Layer ---
